Image classification/requirements.txt
torch
torchvision
scikit-learn
Pillow
PyYAML
tqdm
pandas
matplotlib
seaborn
opencv-python
grad-cam
